Remove circular dependencies, default export
