+++ src/operators/AND/operator.ts
@@ -1,4 +1,5 @@
-import { evaluateArray, getTypeCheckInput } from '../_operatorUtils'
+import { getTypeCheckInput } from '../_operatorUtils'
+import { evaluateArray } from '../../evaluate'
 import { OperatorObject, ParseChildrenMethod, EvaluateMethod } from '../../types'
 import operatorData, { propertyAliases } from './data'
