+++ src/operators/BUILD_OBJECT/operator.ts
@@ -1,4 +1,5 @@
-import { evaluateArray, getTypeCheckInput } from '../_operatorUtils'
+import { getTypeCheckInput } from '../_operatorUtils'
+import { evaluateArray } from '../../evaluate'
 import {
   EvaluatorNode,
   EvaluatorOutput,
