+++ src/operators/COUNT/operator.ts
@@ -1,4 +1,5 @@
-import { evaluateArray, getTypeCheckInput } from '../_operatorUtils'
+import { getTypeCheckInput } from '../_operatorUtils'
+import { evaluateArray } from '../../evaluate'
 import { EvaluateMethod, EvaluatorNode, OperatorObject, ParseChildrenMethod } from '../../types'
 import operatorData, { propertyAliases } from './data'
