+++ src/operators/CUSTOM_FUNCTIONS/operator.ts
@@ -1,5 +1,6 @@
 import extractProperty from 'object-property-extractor'
-import { evaluateArray, getTypeCheckInput } from '../_operatorUtils'
+import { getTypeCheckInput } from '../_operatorUtils'
+import { evaluateArray } from '../../evaluate'
 import { EvaluatorNode, OperatorObject, EvaluateMethod, ParseChildrenMethod } from '../../types'
 import operatorData, { propertyAliases } from './data'
