+++ src/operators/MATCH/operator.ts
@@ -1,5 +1,6 @@
-import { evaluateArray, singleArrayToObject } from '../_operatorUtils'
+import { singleArrayToObject } from '../_operatorUtils'
 import { evaluatorFunction } from '../../evaluate'
+import { evaluateArray } from '../../evaluate'
 import { EvaluatorNode, OperatorObject, EvaluateMethod, ParseChildrenMethod } from '../../types'
 import { isObject, isOperatorNode } from '../../helpers'
 import operatorData, { propertyAliases } from './data'
