+++ src/operators/STRING_SUBSTITUTION/operator.ts
@@ -1,4 +1,5 @@
-import { evaluateArray, zipArraysToObject, getTypeCheckInput } from '../_operatorUtils'
+import { zipArraysToObject, getTypeCheckInput } from '../_operatorUtils'
+import { evaluateArray } from '../../evaluate'
 import { EvaluatorNode, OperatorObject, EvaluateMethod, ParseChildrenMethod } from '../../types'
 import operatorData, { propertyAliases } from './data'
